requirements.txt
yt-dlp==2024.7.9
disnake[voice]==2.9.1
spotipy==2.16.1
python-dotenv==1.0.1
async-timeout==4.0.3
youtube_dl==2021.12.17
psutil==6.0.0
speedtest-cli==2.1.3